aws-s3-django-postgresql-rest/app/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
boto3==1.20.31
botocore==1.23.31
Django==4.0
django-cors-headers==3.10.1
django-storages==1.12.3
djangorestframework==3.13.1
gunicorn==20.1.0
jmespath==0.10.0
Pillow==9.0.0
pkg_resources==0.0.0
psycopg2==2.9.2
psycopg2-binary==2.9.2
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
